+++ pkg/webui/console/components/events/previews/default.js
@@ -12,8 +12,30 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-.overflow-container
-  margin-top: $ls.s * -1
-  display: flex
-  flex: 1
-  overflow-y: auto
+import React from 'react'
+
+import PropTypes from '@ttn-lw/lib/prop-types'
+
+import messages from '../messages'
+
+import DescriptionList from './shared/description-list'
+
+const DefaultPreview = React.memo(({ event }) => {
+  const { identifiers } = event
+
+  if (identifiers && 'device_ids' in identifiers[0]) {
+    return (
+      <DescriptionList>
+        <DescriptionList.Byte title={messages.devAddr} data={identifiers[0].device_ids.dev_addr} />
+      </DescriptionList>
+    )
+  }
+
+  return null
+})
+
+DefaultPreview.propTypes = {
+  event: PropTypes.event.isRequired,
+}
+
+export default DefaultPreview