+++ pkg/webui/console/components/events/previews/error-details.js
@@ -12,8 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-.overflow-container
-  margin-top: $ls.s * -1
-  display: flex
-  flex: 1
-  overflow-y: auto
+import React from 'react'
+
+import ErrorMessage from '@ttn-lw/lib/components/error-message'
+
+import PropTypes from '@ttn-lw/lib/prop-types'
+
+import style from './previews.styl'
+
+const ErrorPreview = React.memo(({ event }) => (
+  <ErrorMessage className={style.plainText} content={event.data} withRootCause />
+))
+
+ErrorPreview.propTypes = {
+  event: PropTypes.event.isRequired,
+}
+
+export default ErrorPreview