+++ pkg/webui/console/components/events/previews/synthetic/default.js
@@ -12,8 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-.overflow-container
-  margin-top: $ls.s * -1
-  display: flex
-  flex: 1
-  overflow-y: auto
+import React from 'react'
+
+import Message from '@ttn-lw/lib/components/message'
+
+import PropTypes from '@ttn-lw/lib/prop-types'
+
+import { eventMessages } from '@console/lib/events/definitions'
+
+import style from '../previews.styl'
+
+const SyntheticEventPreview = React.memo(({ event }) => (
+  <Message className={style.plainText} content={eventMessages[`${event.name}:preview`]} />
+))
+
+SyntheticEventPreview.propTypes = {
+  event: PropTypes.event.isRequired,
+}
+
+export default SyntheticEventPreview