+++ pkg/webui/console/components/events/previews/synthetic/error.js
@@ -12,8 +12,26 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-.overflow-container
-  margin-top: $ls.s * -1
-  display: flex
-  flex: 1
-  overflow-y: auto
+import React from 'react'
+
+import ErrorMessage from '@ttn-lw/lib/components/error-message'
+
+import PropTypes from '@ttn-lw/lib/prop-types'
+
+import { eventMessages } from '@console/lib/events/definitions'
+
+import style from '../previews.styl'
+
+const SyntheticErrorEventPreview = React.memo(({ event }) => (
+  <ErrorMessage
+    className={style.plainText}
+    content={eventMessages[`${event.name}:preview`]}
+    withRootCause
+  />
+))
+
+SyntheticErrorEventPreview.propTypes = {
+  event: PropTypes.event.isRequired,
+}
+
+export default SyntheticErrorEventPreview